+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `verify_nonce()`: `mc_sgx_dcap_types::Quote3` can verify a nonce matches that
   which was provided in a `mc_sgx_core_types::ReportData`.
+- `mc-sgx-dcap-types::QuoteHeader` rust wrapper around `sgx_quote_header_t`
 
 ### Changed